Extract Cache methods to a Contract / Trait (close #57)
