+++ src/Contracts/RepositoryContract.php
@@ -70,54 +70,6 @@ interface RepositoryContract
     public function getModel();
 
     /**
-     * Set the repository cache lifetime.
-     *
-     * @param float|int $cacheLifetime
-     *
-     * @return $this
-     */
-    public function setCacheLifetime($cacheLifetime);
-
-    /**
-     * Get the repository cache lifetime.
-     *
-     * @return float|int
-     */
-    public function getCacheLifetime();
-
-    /**
-     * Set the repository cache driver.
-     *
-     * @param string $cacheDriver
-     *
-     * @return $this
-     */
-    public function setCacheDriver($cacheDriver);
-
-    /**
-     * Get the repository cache driver.
-     *
-     * @return string
-     */
-    public function getCacheDriver();
-
-    /**
-     * Enable repository cache clear.
-     *
-     * @param bool $status
-     *
-     * @return $this
-     */
-    public function enableCacheClear($status);
-
-    /**
-     * Determine if repository cache clear is enabled.
-     *
-     * @return bool
-     */
-    public function isCacheClearEnabled();
-
-    /**
      * Create a new repository model instance.
      *
      * @throws \Rinvex\Repository\Exceptions\RepositoryException
@@ -125,13 +77,6 @@ interface RepositoryContract
      * @return object
      */
     public function createModel();
-
-    /**
-     * Forget the repository cache.
-     *
-     * @return $this
-     */
-    public function forgetCache();
 
     /**
      * Set the relationships that should be eager loaded.
